num_people casted as int

+++ udemy/day_2/4_calculator.py
@@ -6,6 +6,6 @@ perc_tip = input("How much would you like to give? 10, 12 or 15? ")
 num_people = input("How many people to split the bill?")
 
 
-total = round((float(total_bill) * (int(perc_tip)/100) + float(total_bill)) / float(num_people), 2)
+total = round((float(total_bill) * (int(perc_tip)/100) + float(total_bill)) / int(num_people), 2)
 
 print(f"Each person should pay: ${total}")
